Section_5_Django_Framework_avancado/Projeto/requirements.txt
asgiref==3.3.4
coverage==5.5
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.2
django-adminlte2==0.4.3
django-stdimage==5.3.0
gunicorn==20.1.0
model-mommy==2.0.0
Pillow==8.2.0
progressbar2==3.53.1
psycopg2-binary==2.8.6
python-utils==2.5.6
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
